Move CSVLoader to matching hierarchy with core package
